requirements.txt
# Web 框架和相关依赖
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
slowapi==0.1.8

# GitHub 相关
PyGithub==2.1.1

# Notion API
notion-client==2.0.0

# Telegram Bot
python-telegram-bot==20.6

# OpenAI API
openai==1.3.5

# 工具库
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
html2text==2020.1.16